Refactors code.
