+++ tenable/io/exports/api.py
@@ -13,18 +13,27 @@ Methods available on ``tio.exports``:
 '''
 from uuid import UUID
 from json.decoder import JSONDecodeError
+
 from typing_extensions import Literal
 from typing import Dict, Union, List, Optional
 from marshmallow import Schema
 from restfly.errors import RequestConflictError
 from tenable.base.endpoint import APIEndpoint
-from .schema import AssetExportSchema, VulnExportSchema, ComplianceExportSchema, AssetV2ExportSchema
+from .schema import AssetExportSchema, VulnExportSchema, ComplianceExportSchema, AssetV2ExportSchema, WASVulnExportSchema
 from .iterator import ExportsIterator
+
+
+EXPORT_API_PATHS = {
+    'vulns': { '': 'vulns/export'},
+    'assets': { '': 'assets/export', 'v2': 'assets/v2/export' },
+    'compliance': { '': 'compliance/export'},
+    'was/vulns': { 'v1': 'was/v1/export/vulns' }
+}
 
 
 class ExportsAPI(APIEndpoint):
     def _export(self,
-                export_type: Literal['vulns', 'assets', 'compliance'],
+                export_type: Literal['vulns', 'assets', 'compliance', 'was/vulns'],
                 schema: Optional[Schema] = None,
                 use_iterator: bool = True,
                 when_done: bool = False,
@@ -35,7 +44,7 @@ class ExportsAPI(APIEndpoint):
                 **kwargs
                 ) -> Union[ExportsIterator, UUID]:
         '''
-        Get the list of jobs for for the specified datatype.
+        Submit new export job for the specified datatype.
 
         API Documentation for the job listings for
         :devportal:`assets <exports-assets-request-export>`,
@@ -47,16 +56,17 @@ class ExportsAPI(APIEndpoint):
             'vulns': { '': VulnExportSchema },
             'assets': { 'v2': AssetV2ExportSchema, '': AssetExportSchema },
             'compliance': { '': ComplianceExportSchema },
+            'was/vulns': { 'v1': WASVulnExportSchema }
         }
 
         if not schema:
             schema = schemas[export_type][version]()
 
         payload = schema.dump(schema.load(kwargs))
+        path = EXPORT_API_PATHS[export_type][version]
 
         if not export_uuid:
             try:
-                path = f'{export_type}/{version}/export' if version else f'{export_type}/export'
                 export_uuid = self._api.post(path,
                                              json=payload,
                                              box=True
@@ -64,6 +74,7 @@ class ExportsAPI(APIEndpoint):
                 self._log.debug(
                     f'{export_type} {version} export job {export_uuid} initiated'
                 )
+
             except RequestConflictError as conflict:
                 if not adopt_existing:
                     raise conflict
@@ -74,18 +85,22 @@ class ExportsAPI(APIEndpoint):
                     f'Adopted {export_type} {version} export job {export_uuid}.  '
                     f'Original message from platform was "{msg}"'
                 )
+
         if use_iterator:
             return iterator(self._api,
                             type=export_type,
                             uuid=export_uuid,
+                            version=version,
                             _wait_for_complete=when_done,
                             timeout=timeout
                             )
+
         return UUID(export_uuid)
 
     def cancel(self,
-               export_type: Literal['vulns', 'assets', 'compliance'],
+               export_type: Literal['vulns', 'assets', 'compliance', 'was/vulns'],
                export_uuid: UUID,
+               **kwargs
                ) -> str:
         '''
         Cancels the specified export job.
@@ -110,15 +125,18 @@ class ExportsAPI(APIEndpoint):
             >>> tio.exports.cancel('vuln', '{UUID}')
             'CANCELLED'
         '''
-        return self._api.post(f'{export_type}/export/{export_uuid}/cancel',
+        version = '' if export_type == 'assets' else kwargs.pop('version', '')
+        path = EXPORT_API_PATHS[export_type][version]
+        return self._api.post(f'{path}/{export_uuid}/cancel',
                               box=True
                               ).status
 
     def download_chunk(self,
-                       export_type: Literal['vulns', 'assets', 'compliance'],
+                       export_type: Literal['vulns', 'assets', 'compliance', 'was/vulns'],
                        export_uuid: UUID,
                        chunk_id: int,
-                       retries: int = 3
+                       retries: int = 3,
+                       **kwargs
                        ) -> List:
         '''
         Downloads an export chunk from the specified job.
@@ -151,28 +169,31 @@ class ExportsAPI(APIEndpoint):
         downloaded = False
         counter = 0
         resp = []
+        version = '' if export_type == 'assets' else kwargs.pop('version', '')
+        path = EXPORT_API_PATHS[export_type][version]
         while not downloaded and counter <= retries:
             try:
                 resp = self._api.get(
-                    f'{export_type}/export/{export_uuid}/chunks/{chunk_id}'
+                    f'{path}/{export_uuid}/chunks/{chunk_id}'
                 ).json()
                 downloaded = True
             except JSONDecodeError:
                 self._log.warning((
-                    f'{export_type} export {export_uuid} encountered an '
+                    f'{export_type} {version} export {export_uuid} encountered an '
                     f'invalid chunk on chunk id {chunk_id}'
                 ))
                 counter += 1
         if len(resp) < 1:
             self._log.warning((
-                f'{export_type} export {export_uuid} encoundered an empty '
+                f'{export_type} {version} export {export_uuid} encoundered an empty '
                 f'chunk on chunk id {chunk_id}'
             ))
         return resp
 
     def status(self,
-               export_type: Literal['vulns', 'assets', 'compliance'],
+               export_type: Literal['vulns', 'assets', 'compliance', 'was/vulns'],
                export_uuid: UUID,
+               **kwargs
                ) -> Dict:
         '''
         Gets the status of the export job.
@@ -192,12 +213,16 @@ class ExportsAPI(APIEndpoint):
 
             >>> status = tio.exports.status('vulns', '{UUID}')
         '''
-        return self._api.get(f'{export_type}/export/{export_uuid}/status',
+        # Note: Assets export doesn't have v2 api for status call. Its only available for /status call.
+        version = '' if export_type == 'assets' else kwargs.pop('version', '')
+        path = EXPORT_API_PATHS[export_type][version]
+        return self._api.get(f'{path}/{export_uuid}/status',
                              box=True,
                              )
 
     def jobs(self,
-             export_type: Literal['vulns', 'assets'],
+             export_type: Literal['vulns', 'assets', 'was/vulns'],
+             **kwargs
              ) -> Dict:
         '''
         Returns the list of jobs available for a given datatype.
@@ -215,10 +240,12 @@ class ExportsAPI(APIEndpoint):
 
             >>> jobs = tio.exports.jobs('vulns')
         '''
-        return self._api.get(f'{export_type}/export/status', box=True).exports
+        version = '' if export_type == 'assets' else kwargs.pop('version', '')
+        path = EXPORT_API_PATHS[export_type][version]
+        return self._api.get(f'{path}/status', box=True).exports
 
     def initiate_export(self,
-                        export_type: Literal['vulns', 'assets', 'compliance'],
+                        export_type: Literal['vulns', 'assets', 'compliance', 'was/vulns'],
                         **kwargs
                         ):
         """
@@ -654,6 +681,212 @@ class ExportsAPI(APIEndpoint):
             ... )
         '''
         return self._export('vulns', VulnExportSchema(), **kwargs)
+
+    def was_vulns(self, **kwargs) -> Union[ExportsIterator, UUID]:
+        """
+        Initiate a WAS vulnerability export.
+        :devportal:`API Documentation <exports-was-vulns-request-export>`
+        Args:
+            first_found (int, optional):
+                Findings first discovered after this timestamp will be
+                returned.
+            indexed_at (int, optional):
+                Findings indexed into Tenable Vulnerability Management after this timestamp will
+                be returned.
+            last_fixed (int, optional):
+                Findings fixed after this timestamp will be returned.  Note
+                that this filter only applies to fixed data and should not be
+                used when searching for active findings.
+            last_found (int, optional):
+                Findings last observed after this timestamp will be returned.
+            since (int, optional):
+                Findings last observed in any state after this timestamp will
+                be returned.  Cannot be used with ``last_found``,
+                ``first_found``, or ``last_fixed``.
+            plugin_ids (list[int], optional):
+                Only return findings from the specified plugin ids.
+            asset_uuid (list[str], optional):
+                Only return findings for the assets with specific asset-uuids.
+            asset_name (str, optional):
+                Only return findings for the asset with given name.
+            owasp_2010 (list[str], optional):
+                A list of chapters from the OWASP Categories 2010 report for which you want to filter findings returned in the findings export.
+            owasp_2013 (list[str], optional):
+                A list of chapters from the OWASP Categories 2013 report for which you want to filter findings returned in the findings export.
+            owasp_2017 (list[str], optional):
+                A list of chapters from the OWASP Categories 2017 report for which you want to filter findings returned in the findings export.
+            owasp_2021 (list[str], optional):
+                A list of chapters from the OWASP Categories 2021 report for which you want to filter findings returned in the findings export.
+            owasp_api_2019 (list[str], optional):
+                A list of chapters from the OWASP Categories API 2019 report for which you want to filter findings returned in the findings export.
+            severity_modification_type (list[str], optional):
+                Only return vulnerabilities with the specified severity modification type.
+            severity (list[str], optional):
+                Only return findings with the specified severities.
+            state (list[str], optional):
+                Only return findings with the specified states.
+            vpr_score (dict, optional):
+                Only returns findings that meet the specified VPR criteria.
+                The filter is formatted as a dictionary with the mathematical
+                operation as the key.  Supported operations are:
+                .. list-table:: Supported Operations
+                    :widths: auto
+                    :header-rows: 1
+                    * - Operation
+                      - Type
+                      - Description
+                    * - ``eq``
+                      - list[float]
+                      - List of VPR scores that the findings must match.
+                    * - ``neq``
+                      - list[float]
+                      - List of VPR scores that the findings can not match.
+                    * - ``gt``
+                      - float
+                      - VPR scores must be greater than the specified value.
+                    * - ``gte``
+                      - float
+                      - VPR scores must be greater than or equal to the
+                        specified value.
+                    * - ``lt``
+                      - float
+                      - VPR scores must be less than the specified value.
+                    * - ``lte``
+                      - float
+                      - VPR scores must be less than or equal to the
+                        specified value.
+            ipv4s (list[str], optional):
+                Restrict the export to only vulns assigned to assets with specific
+                ip-addresses.
+            include_unlicensed (bool, optional):
+                Should findings for assets that are not licensed be included in
+                the results?
+            num_assets (int, optional):
+                As findings are grouped by asset, how many assets' findings
+                should exist within each data chunk?  If left unspecified the
+                default is ``500``.
+            uuid (str, optional):
+                A predefined export UUID to use for generating an
+                ExportIterator.  Using this parameter will ignore all the
+                filter arguments.
+            use_iterator (bool, optional):
+                Determines if we should return an iterator, or simply the
+                export job UUID.  The default is to return an iterator.
+            when_done (bool, optional):
+                When creating the iterator, setting this flag to true will tell
+                the iterator to wait until the export job has completed before
+                processing the first chunk.  The default behaviour is to start
+                processing chunks of data as soon as they become available.
+            timeout (int, optional):
+                If specified, determines a timeout in seconds to wait for the
+                export job to sit in the queue before cancelling the job and
+                raising a ``TioExportsTimeout`` error.  Once a job has started
+                to be processed, the timeout is ignored.
+            iterator (Iterator, optional):
+                Supports overloading the iterator class to be used to process
+                the datachunks.
+            adopt_existing (bool, optional):
+                Should we automatically adopt an existing Job UUID with we
+                receive a 409 conflict?  Defaults to True.
+        Examples:
+            Examples:
+            Iterating over the results of a WAS vuln export:
+            >>> from tenable.io import TenableIO
+            >>> tio = TenableIO("<apiKey>", "secret")
+            >>> for vuln in tio.exports.was_vulns():
+            ...     print(vuln)
+            Getting findings that have been observed within the last 24 hours
+            >>> import arrow
+            >>> vulns = tio.exports.was_vulns(
+            ...     since=int(arrow.now().shift(days=-1).timestamp())
+            ... )
+            Getting findings that have the ``Region:Chicago`` tag:
+            >>> vulns = tio.exports.was_vulns(
+            ...     tags=[('Region', 'Chicago')]
+            ... )
+        """
+        return self._export('was/vulns', WASVulnExportSchema(), version='v1', **kwargs)
+
+    def was_status(self, export_uuid: UUID, **kwargs) -> Dict:
+        """
+        Gets the status of the WAS export job.
+
+        API Documentation for the status of an export job for the
+        :devportal:`was-vulnerabilities <exports-was-vulns-export-status>` datatypes.
+
+        Args:
+            export_uuid (str):
+                The UUID of the export job.
+
+        Examples:
+
+            >>> status = tio.exports.was_status('{UUID}')
+        """
+        return self.status('was/vulns', export_uuid, version='v1', **kwargs)
+
+    def was_jobs(self, **kwargs) -> Dict:
+        """
+        Returns the list of WAS export jobs.
+
+        API Documentation for the job listing APIs for
+        :devportal:`WAS vulnerabilities <exports-was-vulns-export-status-recent>`
+        datatypes.
+
+        Examples:
+
+            >>> jobs = tio.exports.was_jobs()
+        """
+        return self.jobs('was/vulns', version='v1', **kwargs)
+
+    def cancel_was_job(self, export_uuid: UUID, **kwargs) -> str:
+        """
+        Cancels the specified WAS export job.
+
+        API Documentation for cancel export jobs with
+        :devportal:`was vulnerabilities <exports-was-vulns-export-cancel>` datatype.
+
+        Args:
+            export_uuid:
+                The export job's unique identifier.
+
+        Returns:
+            str:
+                The status of the job.
+
+        Example:
+
+            >>> tio.exports.cancel_was_job('{UUID}')
+            'CANCELLED'
+        """
+        return self.cancel('was/vulns', export_uuid, version='v1', **kwargs)
+
+    def download_was_chunk(self,
+                       export_uuid: UUID,
+                       chunk_id: int,
+                       retries: int = 3,
+                       **kwargs
+                       ) -> List:
+        """
+        Downloads a WAS export chunk from the specified job.
+
+        API Documentation for downloading an export chunk for
+        :devportal:`WAS vulnerabilities <exports-was-vulns-download-chunk>`.
+
+        Args:
+            export_uuid:
+                The export job's unique identifier.
+            chunk_id:
+                The identifier for the specific chunk to download.
+
+        Returns:
+            List:
+                The list of objects that entail the chunk of data requested.
+
+        Example:
+
+            >>> chunk = tio.exports.download_chunk('{UUID}', 1)
+        """
+        return self.download_chunk('was/vulns', export_uuid, chunk_id, retries, version='v1', **kwargs)
 
     def list_compliance_export_jobs(self):
         """
