+++ tenable/io/exports/iterator.py
@@ -61,6 +61,7 @@ class ExportsIterator(APIIterator):  # noqa: PLR0902
     timeout: int = None
     uuid: str
     type: str
+    version: str
     status: str
     start_time: int
     count: int = 0
@@ -74,7 +75,7 @@ class ExportsIterator(APIIterator):  # noqa: PLR0902
         super().__init__(api, **kwargs)
 
     def _get_status_resp(self) -> Dict:
-        return self._api.exports.status(self.type, self.uuid)
+        return self._api.exports.status(self.type, self.uuid, version=self.version)
 
     def _get_status(self) -> Dict:
         '''
@@ -82,8 +83,9 @@ class ExportsIterator(APIIterator):  # noqa: PLR0902
         iterator is within the export job and return the status to the caller
         '''
         status = self._get_status_resp()
-        self._log.debug('%s export %s is currently %s',
+        self._log.debug('%s export %s %s is currently %s',
                         self.type,
+                        self.version,
                         self.uuid,
                         status.status
                         )
@@ -155,7 +157,7 @@ class ExportsIterator(APIIterator):  # noqa: PLR0902
         return self.chunks
 
     def _get_chunk_data(self, type, uuid, chunk_id) -> List:
-        return self._api.exports.download_chunk(type, uuid, chunk_id)
+        return self._api.exports.download_chunk(type, uuid, chunk_id, version=self.version)
 
     def _get_page(self):
         '''
@@ -214,7 +216,7 @@ class ExportsIterator(APIIterator):  # noqa: PLR0902
         '''
         Cancels the current export
         '''
-        self._api.exports.cancel(self.type, self.uuid)
+        self._api.exports.cancel(self.type, self.uuid, version=self.version)
 
     def run_threaded(self,
                      func: Any,
@@ -230,6 +232,7 @@ class ExportsIterator(APIIterator):  # noqa: PLR0902
             * export_uuid (str): Receiver for the export job UUID.
             * export_type (str): Receiver for the export data-type.
             * export_chunk_id (int): Receiver for the export chunk id.
+            * version (int): Receiver for the export version.
 
         Args:
             func:
@@ -248,9 +251,10 @@ class ExportsIterator(APIIterator):  # noqa: PLR0902
             >>> def write_chunk(data,
             ...                 export_uuid: str,
             ...                 export_type: str,
-            ...                 export_chunk_id: int
+            ...                 export_chunk_id: int,
+            ...                 version: int
             ...                 ):
-            ...     fn = f'{export_type}-{export_uuid}-{export_chunk_id}.json'
+            ...     fn = f'{export_type}-{export_uuid}-{export_chunk_id}-{version}.json'
             ...     with open(fn, 'w') as fobj:
             ...         json.dump(data, fobj)
             >>>
@@ -280,6 +284,7 @@ class ExportsIterator(APIIterator):  # noqa: PLR0902
             kw['export_uuid'] = self.uuid
             kw['export_type'] = self.type
             kw['export_chunk_id'] = chunk_id
+            kw['version'] = self.version
             self._log.debug(
                 (f'{self.type} export {self.uuid} chunk {chunk_id} '
                  'has been downloaded and the data has been handed '
